osspeak/requirements.txt
altgraph==0.16.1
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
future==0.16.0
idna==2.7
importlib-metadata==0.19
keyboard==0.13.2
lark-parser==0.8.0
macholib==1.11
more-itertools==4.3.0
mouse==0.7.0
packaging==19.1
pefile==2018.8.8
pluggy==0.12.0
py==1.6.0
PyInstaller==3.5
pyparsing==2.4.2
pypiwin32==223
pytest==5.1.1
pywin32==224
pywin32-ctypes==0.2.0
pywindow==0.0.3
requests==2.19.1
screengrid==0.0.11
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
websockets==6.0
zipp==0.6.0
